requirements.txt
Django==1.4.1
psycopg2==2.4.5
dj-database-url==0.2.0
tweepy==1.11
gunicorn==0.14.6
django-storages==1.1.5
South==0.7.6
django-twilio==0.4
django-dajaxice==0.5.4
pillow==1.7.8
newrelic==1.9.0.21
pubnub==3.3
